feat(core,membrane,harness): extend elimination backoff, add Clifford algebra runtime math, fix fault injection
Expand elimination backoff implementation in core allocator, add Clifford
algebra module to membrane runtime_math, refine ptr_validator, and fix
fault injection harness.

+++ crates/frankenlibc-harness/src/fault_injection.rs
@@ -534,7 +534,7 @@ fn execute_memory_case(
                         scenario: scenario.id.clone(),
                         reason: format!("failed to allocate {size} bytes"),
                     })?;
-            if !pipeline.inject_trailing_canary_corruption(ptr, size, fill_byte) {
+            if !pipeline.inject_trailing_canary_corruption(ptr as usize, size, fill_byte) {
                 return Err(FaultInjectionError::ExecutionFailure {
                     scenario: scenario.id.clone(),
                     reason: "failed to inject canary corruption".to_string(),
